Move jmespath.go/jmespath to jmespath.go
This removes confusion with not being able to go get the
project library from the project's root, as per review feedback.
