+++ jputil/util.go
