+++ lexer.go
