+++ lexer_test.go
