+++ toktype_string.go
